requirements.txt
channels==3.0.5
Django==5.1.6
django_channels==0.7.0
djangorestframework==3.15.2
python-decouple==3.8
six==1.17.0
